More pathfinder
Made a horrible method for getting an arraylist of edge tiles in an
appropreate order. she'll be right.

Going off the assumption findEdgeTiles works, the looping through thing
should be all good, including breakpoint as soon as it finds a clear
route, and making sure it finds shorter routes first.